at create-post & text-overflow in post

+++ components/Post.php
@@ -1,32 +1,40 @@
 <?php
 $Post = function ($post_id) {
-    $db = import('./Database/db');
-    $db->query("UPDATE post SET `post_view`=`post_view`+1 WHERE post_id = $post_id");
+    $db = new Database;
+    $db->postView_UP($post_id);
+
     if (isset($_POST['like'])) {
         if (!isset($_SESSION['usr'])) {
             header("Location: /login");
             die;
         }
+        $usr_id = (int) $_SESSION['usr'];
+        $post_id = (int) $_POST['post_id'] ?? 0;
 
-        $db->query("INSERT INTO `post_like`
-    (`pl_id`, `post_id`, `usr_id`) VALUES 
-    (NULL, {$_POST['post_id']}, {$_SESSION['usr']})");
+        if ($db->isLikePost($post_id, $usr_id)) {
+            $db->deleteLikePost($post_id, $usr_id);
+        } else {
+            $db->insertLikePost($post_id, $usr_id);
+        }
         header("Refresh:0");
         die;
     }
 
     if (isset($_POST['deletePost'])) {
-        $getPost = fetch($db->query("SELECT * FROM post WHERE post_id = {$_POST['post_id']}"));
-        unlink('./public/posts/' . $getPost['post_img']);
-        $db->query("DELETE FROM post WHERE post_id = {$_POST['post_id']}");
-        header("Refresh:0");
-        die;
+        $post_id = $_POST['post_id'] ?? 0;
+        $getPost = $db->getPost_ByID($post_id);
+        if (!isset($_SESSION['usr']) || $_SESSION['usr'] != $getPost['post_usr_id'] && (!isset($_SESSION['status']) || $_SESSION['status'] != 'admin')) {
+        } else {
+            unlink('./public/posts/' . $getPost['post_img']);
+            $db->deletePost_ByID($post_id);
+            header("Refresh:0");
+            die;
+        }
     }
 
-    $allPost = $db->query("SELECT * FROM post WHERE post_id = $post_id");
-    $post = fetch($allPost);
-    $usr_post = fetch($db->query("SELECT * FROM usr WHERE usr_id = {$post['post_usr_id']}"));
-    $cat = fetch($db->query("SELECT * FROM cat WHERE cat_id = {$post['post_cat_id']}"));
+    $post = $db->getPost_ByID($post_id);
+    $usr_post = $db->getUser_ByID($post['post_usr_id']);
+    $cat = $db->getCate_ByID($post['post_cat_id']);
 ?>
 
 
@@ -47,7 +55,7 @@ $Post = function ($post_id) {
             </div>
             <div id="dropdownpost" class="z-30 hidden text-base list-none bg-white divide-y divide-gray-100 rounded shadow w-44 dark:bg-gray-700">
                 <form method="post">
-                    <input type="hidden" name="post_id" value="<?php echo $post['post_id']; ?>">
+                    <input type="hidden" name="post_id" value="<?php echo $post['post_id'] ?? 0; ?>">
                     <ul class="py-1" aria-labelledby="dropdownPost">
                         <li>
                             <button name="deletePost" class="text-rose-600 block w-full text-left px-4 py-2 text-sm hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-200 dark:hover:text-white">ลบ</button>
@@ -60,25 +68,25 @@ $Post = function ($post_id) {
 
         <div class="flex items-center px-3">
             <div>
-                <img class="w-9 rounded-full inline-block" src="/public/profile/<?php echo $usr_post['usr_img'] ?? ""; ?>" alt="profile image">
+                <img class="w-9 h-9 rounded-full object-cover inline-block" src="/public/profile/<?php echo $usr_post['usr_img'] ?? ""; ?>" onerror="this.onerror=null; this.src='/public/default/profile.png'" alt="profile image">
             </div>
             <div class="px-3">
                 <a class="hover:underline" href="/<?php echo $usr_post['usr_username'] ?? ""; ?>"><?php echo $usr_post['usr_name'] ?? ""; ?></a>
-                <div class="text-gray-500 text-sm">โพสต์เมื่อ <?php echo $post['post_date'] ?? ""; ?> ⦁ <?php echo $cat['cat_name'] ?? ""; ?></div>
+                <div class="text-gray-500 text-sm">โพสต์เมื่อ <?php echo $post['post_date'] ?? ""; ?> ⦁ <?php echo $cat['cat_name'] ?? "ไม่พบหมวดหมู่"; ?></div>
             </div>
         </div>
-        <div class="my-3 px-3 overflow-hidden ">
-            <?php echo str_replace("\n", '<br>', $post['post_detail']) ?? ""; ?>
+        <div class="my-3 px-3 ">
+            <?php echo htmlchar($post['post_detail']) ?? ""; ?>
         </div>
         <div>
-            <img class="w-full" src="/public/posts/<?php echo $post['post_img'] ?? ""; ?>" alt="image post">
+            <img class="w-full" src="/public/posts/<?php echo $post['post_img'] ?? ""; ?>" onerror="this.onerror=null; this.src='/public/default/post.png'" alt="image post">
         </div>
         <div class="m-3">
             <span><img class="inline-block w-6" src="/public/icons/f-heart.svg" alt="full heart icon">
-                <?php echo $db->query("SELECT * FROM post_like WHERE post_id = {$post['post_id']}")->num_rows ?? 0; ?>
+                <?php echo $db->getPostLikeCount($post['post_id']); ?>
             </span>
             <span class="float-right">
-                <?php echo $db->query("SELECT * FROM post_detail WHERE post_id = {$post['post_id']}")->num_rows ?? 0; ?>
+                <?php echo $db->getPostCommentCount($post['post_id']) ?? 0; ?>
                 ความคิดเห็น
                 <span class="mx-3">
                     <?php echo $post['post_view'] ?? 0; ?> รับชม
@@ -88,7 +96,7 @@ $Post = function ($post_id) {
         <hr class="border">
         <div class="p-3 pb-0 text-gray-600">
             <form method="post">
-                <?php if (!isset($_SESSION['usr']) || $db->query("SELECT * FROM post_like WHERE post_id = {$post['post_id']} AND usr_id = {$_SESSION['usr']}")->num_rows == 0) : ?>
+                <?php if (!isset($_SESSION['usr']) || !$db->isLikePost($post['post_id'], $_SESSION['usr'])) : ?>
                     <input type="hidden" name="post_id" value="<?php echo $post['post_id']; ?>">
                     <button name="like" class="py-1 px-3 hover:bg-gray-200 rounded-lg"><img class="inline-block w-6" src="/public/icons/heart.svg" alt="heart icon"> ถูกใจ</button>
                 <?php else : ?>
